:sparkles: show correct home link url
