+++ src/main/java/com/example/project/controller/HomeController.java
@@ -58,117 +58,5 @@ public class HomeController {
         return "dashboard";
     }
 
-    @GetMapping("/tutor")
-    public String tutorPage(@AuthenticationPrincipal UserDetails userDetails, Model model) {
-        User tutor = userService.findByUsername(userDetails.getUsername());
-
-        if (tutor.getRole() != 2) { // Ensure tutor role
-            return "redirect:/access-denied";
-        }
-
-        Long tutorId = tutor.getId();
-        LocalDate today = LocalDate.now();
-
-        // Get all sessions for this tutor
-        List<Session> sessions = sessionService.getTutorSessions(tutorId, null);
-
-        // Upcoming sessions: today or later
-        List<Session> upcomingSessions = sessions.stream()
-                .filter(s -> !s.getDate().isBefore(today))
-                .sorted(Comparator.comparing(Session::getDate).thenComparing(Session::getTimeSlot))
-                .collect(Collectors.toList());
-
-        // Past sessions: before today, only where studentId is not null
-        List<Session> pastSessions = sessions.stream()
-                .filter(s -> s.getDate().isBefore(today) && s.getStudentId() != null)
-                .sorted(Comparator.comparing(Session::getDate).reversed())
-                .collect(Collectors.toList());
-
-        Map<Long, String> studentNames = new HashMap<>();
-        for (Session s : sessions) {
-            if (s.getStudentId() != null) {
-                User student = userService.findById(s.getStudentId());
-                Long sId = s.getStudentId();
-                studentNames.put(sId, student.getFirstName() + " " + student.getLastName());
-            }
-        }
-
-        model.addAttribute("upcomingSessions", upcomingSessions);
-        model.addAttribute("pastSessions", pastSessions);
-        model.addAttribute("studentNames", studentNames);
-        model.addAttribute("tutorId", tutorId);
-
-        return "tutor";
-    }
-
-    /**
-     * Add a new session (Tutor)
-     */
-    @PostMapping("/tutor/session")
-    public String createSession(@AuthenticationPrincipal UserDetails userDetails,
-                                @RequestParam("date") String date,
-                                @RequestParam("timeSlot") String timeSlot,
-                                RedirectAttributes redirectAttributes) {
-        try {
-            User tutor = userService.findByUsername(userDetails.getUsername());
-            if (tutor.getRole() != 2) {
-                return "redirect:/access-denied";
-            }
-            LocalDate sessionDate = LocalDate.parse(date);
-            sessionService.createSession(tutor.getId(), sessionDate, timeSlot);
-            redirectAttributes.addFlashAttribute("success", "Session created successfully!");
-        } catch (RuntimeException e) {
-            redirectAttributes.addFlashAttribute("error", e.getMessage());
-        }
-        return "redirect:/tutor";
-    }
-
-    /**
-     * Delete a session (Tutor) — only for upcoming sessions
-     */
-    @PostMapping("/tutor/session/delete")
-    public String deleteSession(@AuthenticationPrincipal UserDetails userDetails,
-                                @RequestParam("sessionId") Long sessionId, RedirectAttributes redirectAttributes) {
-        User tutor = userService.findByUsername(userDetails.getUsername());
-
-        if (tutor.getRole() != 2) {
-            return "redirect:/access-denied";
-        }
-
-        // Verify session is upcoming before deleting
-        Session session = sessionService.getTutorSessions(tutor.getId(), null).stream()
-                .filter(s -> s.getId().equals(sessionId))
-                .findFirst()
-                .orElseThrow(() -> new RuntimeException("Session not found"));
-
-
-        if (!session.getDate().isBefore(LocalDate.now())) {
-            sessionService.deleteSession(sessionId);
-            redirectAttributes.addFlashAttribute("success", "Session deleted successfully!");
-        } else {
-            redirectAttributes.addFlashAttribute("error", "Cannot delete past sessions.");
-        }
-
-        return "redirect:/tutor";
-    }
-
-    @GetMapping("/tutor/profile")
-    public String showTutorProfile(@AuthenticationPrincipal UserDetails userDetails, Model model) {
-        User user = userService.findByUsername(userDetails.getUsername());
-
-        // Optional: double-check in controller level if needed
-        if (user.getRole() != 2) { // not a tutor
-            return "redirect:/access-denied"; // make this page if needed
-        }
-
-        model.addAttribute("user", user);
-        return "tutor-profile";
-    }
-
-    @PostMapping("/tutor/profile")
-    public String updateTutorProfile(@AuthenticationPrincipal UserDetails userDetails,
-                                     @ModelAttribute("user") User updatedUser) {
-        userService.updateUserProfile(userDetails.getUsername(), updatedUser);
-        return "redirect:/tutor/profile";
-    }
+    
 }
